refactor: move actions module, add derives & docs
